chore: regenerate types

+++ tests/src/interfaces/default/types.ts
@@ -1303,7 +1303,7 @@ export interface PalletConfigurationAppPromotionConfiguration extends Struct {
 export interface PalletConfigurationCall extends Enum {
   readonly isSetWeightToFeeCoefficientOverride: boolean;
   readonly asSetWeightToFeeCoefficientOverride: {
-    readonly coeff: Option<u32>;
+    readonly coeff: Option<u64>;
   } & Struct;
   readonly isSetMinGasPriceOverride: boolean;
   readonly asSetMinGasPriceOverride: {
